Add sitemap for google
We cannot use a handy plugin like https://github.com/jekyll/jekyll-sitemap here because of the "special" way the collection works (it contains dupliacte content for some reason). So we need to build it manually.

The posts part is actually empty.

The pages show the content pages. For some reason we need to manually exclude some static files here. Might be a missing config somewhere.

The haushalte list all haushalte but switch the  changefreq based on their content – more often if they have content. This is just for fun, google will to what it wants anyway.

The page/bundesland/* are excluded here since their layout seems to be broken.