Update Jekyll to 3.7.2 and add redirect-from-plugin
The 'jekyll-redirect-from' requires a new jekyll. And its a good idea to update anyways. I did not see any issues with it as well.

As part of the update the letters=[] needs to be removed, it causes liquid-warnings while jekyll generates. Also, its an empty variable that should be removed in any case.