Potsdam: Improve presentation, add translations
Add economic classification as second dimension.

Remove empty filters "". These generally don't make sense, because
numbers from e.g. different years would be accumulated.

Provide German translations for labels.

Tried to provide the config for a barchart as in Berlin, but nothing is
displayed. So disabled it again for now.